+++ algorithms.hpp
@@ -17,7 +17,7 @@ public:
     void GreedyColoring(Graph *g);
     int executeRandomGreedy(Graph *g, int n);
     int GreedyColoringCustom(Graph *, std::vector<int>*);
-    long calcTriangles(int, arma::Mat<int>*);
+    long calcTriangles(int, arma::Mat<short>*);
     long calcTriangles2(int numKnot, std::vector<std::set<int>>* neighbours);
 
 private:
